Support `include` and `panel` layout
Includes should be loaded okay, but saving assumes `filename` is of the
form `skin_INCLUDE.xml` (possibly with a path).

Panels were supported as an element (having a name), but not as a
container (position and size, with the panel name as a child).

Restrict search results to as many will fit.